requirements.txt
python-dotenv
discord
requests
pyxivapi
asyncio